update cron-task delete job code

+++ internal/computing/cron_task.go
@@ -196,7 +196,18 @@ func (task *CronTask) watchExpiredTask() {
 				// compatible with space_uuid
 				spaceUuidDeployName := constants.K8S_DEPLOY_NAME_PREFIX + strings.ToLower(job.SpaceUuid)
 				if _, ok = deployOnK8s[spaceUuidDeployName]; ok {
-					if NewJobService().GetJobEntityBySpaceUuid(job.SpaceUuid) > 0 {
+					if NewJobService().GetJobEntityBySpaceUuid(job.SpaceUuid) > 0 && time.Now().Unix() < job.ExpireTime {
+						if job.Status != models.JOB_RUNNING_STATUS {
+							foundDeployment, err := NewK8sService().k8sClient.AppsV1().Deployments(job.NameSpace).Get(context.TODO(), job.K8sDeployName, metav1.GetOptions{})
+							if err != nil {
+								continue
+							}
+							if foundDeployment.Status.AvailableReplicas > 0 {
+								job.PodStatus = models.POD_RUNNING_STATUS
+								job.Status = models.JOB_RUNNING_STATUS
+								NewJobService().UpdateJobEntityByJobUuid(job)
+							}
+						}
 						continue
 					}
 
@@ -221,13 +232,13 @@ func (task *CronTask) watchExpiredTask() {
 			currentTime := time.Now()
 			createdTime := time.Unix(job.CreateTime, 0)
 			createDuration := currentTime.Sub(createdTime)
-			if createDuration.Hours() <= 2 && job.Status != models.JOB_RUNNING_STATUS {
-				continue
-			}
 
 			if job.NameSpace != "" && job.K8sDeployName != "" {
 				foundDeployment, err := NewK8sService().k8sClient.AppsV1().Deployments(job.NameSpace).Get(context.TODO(), job.K8sDeployName, metav1.GetOptions{})
 				if err != nil {
+					if createDuration.Hours() <= 2 && job.Status != models.JOB_RUNNING_STATUS {
+						continue
+					}
 					if errors.IsNotFound(err) {
 						// delete job
 						logs.GetLogger().Warnf("not found deployment on the cluster, job_uuid: %s, deployment: %s", job.JobUuid, job.K8sDeployName)
@@ -238,7 +249,7 @@ func (task *CronTask) watchExpiredTask() {
 					continue
 				}
 
-				if foundDeployment.Status.AvailableReplicas == 0 { // need to delete
+				if foundDeployment.Status.AvailableReplicas == 0 && createDuration.Hours() > 2 { // need to delete
 					DeleteJob(job.NameSpace, job.JobUuid, "cron-task correction status")
 					deleteSpaceIdAndJobUuid[job.JobUuid] = job.SpaceUuid + "_" + job.JobUuid
 					continue
@@ -466,6 +477,10 @@ func addNodeLabel() {
 			}
 		}
 	}
+}
+
+func changeJobStatus() {
+
 }
 
 func reportJobStatus(jobUuid string, deployStatus int) bool {
